metrics_loaders/cypress/mochawesome/requirements.txt
decorator==5.1.1
jsonpath-rw==1.4.0
jsonpath-rw-ext==1.2.2
mysql-connector-python==8.0.31
pbr==5.11.0
ply==3.11
protobuf==3.20.1
six==1.16.0
toml==0.10.2
